Move Camera implementation from world to engine package
